tests: fix shader_jit_compiler and add pica_float
The Uniform Read and Address Register Offset ran into precision
differences between f32 and f24, which can be easily fixed and
are not the point of the tests.
As for the LG2 and EX2 tests that were failing, they were wrong.
While it is true that 2^79.7 ~= 1e24, the value is bigger than the
biggest representable value in f24, therefore both EX2 and LG2(?)
should result in Inf.