Android: Implement quick configure for external controllers
This implements a fast-button-mapper.

The user can now click the Auto Configuration setting in the gamepad-section, which will open a new dialog.

This dialog will then iterate over the available buttons, where the user can quickly press a button on the gamepad without having to pre-select which button this press should be assigned to.

Ported from Lime3DS.